make level_of_politics tinyint

+++ api/database/sql/V12__create_level_of_politics.sql
@@ -1,5 +1,5 @@
 create table level_of_politics (
-    id int not null auto_increment,
+    id tinyint not null auto_increment,
     name varchar(128) not null,
     description text,
     primary key (id),
